Update mock observability extensions to the new model
