+++ tests/jballerina-integration-test/src/test/resources/observability/metrics_tests/ballerina.conf
@@ -1,3 +1,3 @@
 [b7a.observability.metrics]
 enabled=true    # Flag to enable Metrics
-reporter="BMockMetricsReporter"
+reporter="mock"
