+++ tests/jballerina-integration-test/src/test/resources/observability/tracing_tests/ballerina.conf
@@ -1,3 +1,3 @@
 [b7a.observability.tracing]
 enabled=true    # Flag to enable Tracing
-name="BMockTracer"
+name="mock"
