+++ tests/observability-test-utils/src/main/java/module-info.java
@@ -6,4 +6,5 @@ module io.ballerina.observability.test.utils {
     requires io.netty.buffer;
     requires io.netty.codec.http;
     requires io.ballerina.lang;
+    requires opentracing.api;
 }
