+++ tests/observability-test-utils/src/main/java/org/ballerina/testobserve/Constants.java
@@ -1,0 +1,27 @@
+/*
+ * Copyright (c) 2020, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ * WSO2 Inc. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package org.ballerina.testobserve;
+
+import io.ballerina.runtime.api.Module;
+
+/**
+ * Constants related to Ballerina Observe Module.
+ */
+public class Constants {
+    public static final Module OBSERVE_MODULE = new Module("ballerina", "testobserve", "0.0.0");
+}
