+++ tests/observability-test-utils/src/main/java/org/ballerina/testobserve/tracing/extension/MockTracerUtils.java
@@ -32,7 +32,7 @@ import java.util.stream.Collectors;
  */
 public class MockTracerUtils {
     public static Object getFinishedSpans(String serviceName) {
-        MockTracer mockTracer = BMockTracer.getTracerMap().get(serviceName);
+        MockTracer mockTracer = BMockTracerProviderFactory.BMockTracerProvider.getTracerMap().get(serviceName);
         List<BMockSpan> mockSpans;
         if (mockTracer == null) {
             mockSpans = Collections.emptyList();
