--- tests/observability-test-utils/src/main/resources/META-INF/services/io.ballerina.runtime.observability.metrics.spi.MetricReporter
@@ -1,1 +1,0 @@
-org.ballerina.testobserve.metrics.extension.BMockMetricsReporter
