--- tests/observability-test-utils/src/main/resources/META-INF/services/io.ballerina.runtime.observability.tracer.OpenTracer
@@ -1,1 +1,0 @@
-org.ballerina.testobserve.tracing.extension.BMockTracer
